feat: configure storybook
