+++ .storybook/preview.tsx
@@ -1,0 +1,28 @@
+import type { Preview } from '@storybook/react';
+import React from 'react';
+
+import { ThemeProvider } from '@/components/theme-provider';
+import '@/components/theme-provider/styles.css';
+
+const preview: Preview = {
+  parameters: {
+    actions: {
+      argTypesRegex: '^on[A-Z].*',
+    },
+    controls: {
+      matchers: {
+        color: /(background|color)$/i,
+        date: /Date$/i,
+      },
+    },
+  },
+  decorators: [
+    (Story) => (
+      <ThemeProvider>
+        <Story />
+      </ThemeProvider>
+    ),
+  ],
+};
+
+export default preview;
