+++ app/layout.tsx
@@ -2,6 +2,7 @@ import type { Metadata } from 'next';
 import type { PropsWithChildren } from 'react';
 
 import { ThemeProvider } from '@/components/theme-provider';
+import '@/components/theme-provider/styles.css';
 
 export const metadata: Metadata = {
   description: 'Welcome, welcome, welcome!',
