+++ components/theme-provider/radix-themes.provider.tsx
@@ -1,7 +1,6 @@
 import type { PropsWithChildren } from 'react';
 
 import { Theme, ThemePanel } from '@radix-ui/themes';
-import '@radix-ui/themes/styles.css';
 
 export function RadixThemesProvider({ children }: PropsWithChildren) {
   return (
