+++ components/theme-provider/styles.css
@@ -1,0 +1,1 @@
+@import '@radix-ui/themes/styles.css';
