+++ package.json
@@ -26,7 +26,6 @@
     "@storybook/addon-essentials": "^7.6.4",
     "@storybook/addon-interactions": "^7.6.4",
     "@storybook/addon-links": "^7.6.4",
-    "@storybook/addon-onboarding": "^1.0.9",
     "@storybook/blocks": "^7.6.4",
     "@storybook/nextjs": "^7.6.4",
     "@storybook/react": "^7.6.4",
